requirements.txt
SpeechRecognition==3.8.1
python-docx==0.8.11
PyMuPDF==1.21.1
Pillow==9.5.0
pytesseract==0.3.10
pdf2image==1.16.3
transformers==4.29.2
nltk==3.8.1
Spire.Xls
